+++ src/McDataTable.vue
@@ -73,8 +73,8 @@ const props = withDefaults(defineProps<DataTablePropsType>(), {
   },
   sortStyle              : () => {
     return {
-      asc : "fa fa-caret-up",
-      desc: "fa fa-caret-down",
+      asc : "keyboard_arrow_up",
+      desc: "keyboard_arrow_down",
     }
   },
   permittedEvents        : () => ["click", "mouseover", "mouseleave", "mouseenter"],
@@ -185,5 +185,3 @@ provide("mcSetSearchItemsCount", setSearchItemsCount)
 <style scoped>
 
 </style>
-
-