+++ src/McExplorerDataTable.vue
@@ -74,8 +74,8 @@ const props = withDefaults(defineProps<DataTablePropsType>(), {
   },
   sortStyle              : () => {
     return {
-      asc : "fa fa-caret-up",
-      desc: "fa fa-caret-down",
+      asc : "keyboard_arrow_up",
+      desc: "keyboard_arrow_down",
     }
   },
   permittedEvents        : () => ["click", "mouseover", "mouseleave", "mouseenter"],